feat(tracking): add tracking of smartlink event to CreateSmartLinks
- small refactors + test on tracking called when creating smart links
