+++ store/src/main/java/com/zextras/mailbox/tracking/Tracking.java
@@ -6,16 +6,12 @@ package com.zextras.mailbox.tracking;
 
 import io.vavr.control.Try;
 
+/**
+ * Tracks events/actions performed by the user to use as metrics/insights on Carbonio usages.
+ * Tracking of data is anonymous and not linkable to user information/personal data.
+ *
+ */
 public interface Tracking {
-
-  // curl -i "https://analytics.zextras.tools/matomo.php?
-  // idsite=7 -> TODO: is idsite always the same? for now let's use a static value
-  // &rec=1&send_image=0&apiv=1 -> these params are static
-
-  // &e_c=Mail -> Category
-  // &e_a=SendEmailWithSmartLinks&  -> Action
-  // uid=UserA" -> user
-  // -> these are dynamic
 
   Try<Void> sendEvent(Event event);
