+++ store/src/main/java/com/zimbra/cs/service/mail/MailService.java
@@ -9,6 +9,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.zextras.carbonio.files.FilesClient;
 import com.zextras.files.client.GraphQLFilesClient;
 import com.zextras.mailbox.smartlinks.FilesSmartLinksGenerator;
+import com.zextras.mailbox.tracking.MatomoTracking;
+import com.zextras.mailbox.tracking.Tracking;
 import com.zimbra.common.soap.MailConstants;
 import com.zimbra.cs.service.MailboxAttachmentService;
 import com.zimbra.soap.DocumentDispatcher;
@@ -237,10 +239,14 @@ public class MailService implements DocumentService {
     dispatcher.registerHandler(
         QName.get("CreateSmartLinksRequest", MailConstants.NAMESPACE),
         new CreateSmartLinks(new FilesSmartLinksGenerator(
-            new GraphQLFilesClient(getFilesClient(), new ObjectMapper()),
-            filesCopyHandler)
+            new GraphQLFilesClient(getFilesClient(), new ObjectMapper()), filesCopyHandler),
+            getTracking()
         )
     );
+  }
+
+  protected Tracking getTracking() {
+    return new MatomoTracking("https://analytics.zextras.tools");
   }
 
   protected FilesClient getFilesClient() {
